以项目纬度重构

+++ src/main/java/cn/wzy/dao/IProjectInfoMapper.java
@@ -20,5 +20,5 @@ public interface IProjectInfoMapper extends BaseMapper<ProjectInfo> {
      * @param id
      * @return
      */
-    List<ProjectInfo> getByUserId(Long id);
+    List<ProjectInfo> getByUserId(Integer id);
 }
